lib/js/datachannel_test_2peer/label/003.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - Label Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">            
            var label = "-", size = 15;
            var dataChannel = null;
            for (var i = 0; i < size; i++) {
		        label += label;
		    }
            var test = async_test("Set up a DataChannel with a label of length " + label.length / 1024 + " KB - check label");
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label);
                    dataChannel.onopen = test.step_func(function() {
                        assert_equals(dataChannel.label.length, label.length, "Wrong label length ");
                        test.done();
                    });
                });
            } else {
						peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel = e.channel;
                    dataChannel.onopen = test.step_func(function() {
                        assert_equals(dataChannel.label.length, label.length, "Wrong label.length");
                        test.done();
                    });
                });
            }

		</script>
	</body>
</html>
